+++ magicbell/client.py
@@ -2,7 +2,7 @@ import typing
 
 import httpx
 
-from .api import ProjectAPI, RealtimeAPI, UserAPI
+from .api import ChannelsAPI, ProjectAPI, RealtimeAPI, UserAPI
 from .configuration import Configuration
 
 
@@ -25,6 +25,7 @@ class MagicBell:
     http_client: httpx.AsyncClient
     projects: ProjectAPI
     realtime: RealtimeAPI
+    channels: ChannelsAPI
 
     def __init__(
         self,
@@ -51,6 +52,7 @@ class MagicBell:
         self.projects = ProjectAPI(self.http_client, self.configuration)
         self.realtime = RealtimeAPI(self.http_client, self.configuration)
         self.users = UserAPI(self.http_client, self.configuration)
+        self.channels = ChannelsAPI(self.http_client, self.configuration)
 
     async def __aenter__(self):
         if not self._is_unmanaged_http_client: