+++ CHANGELOG.md
@@ -19,3 +19,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Create project
   - Update project
   - Delete project
+  - Updating channels for a project